app/requirements.txt
torch
transformers
huggingface_hub
gradio
Pillow
